+++ parser/internal/names/resolver_test.go
